CLIENT-3268 Rename MRT to transaction in docs, comments and error strings. Rename MRT to TXN in private/protected variables. Rename "multi-record transaction" to "transaction" in most cases, except the Txn class description.

+++ client/src/com/aerospike/client/command/Buffer.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2012-2024 Aerospike, Inc.
+ * Copyright 2012-2025 Aerospike, Inc.
  *
  * Portions may be licensed to Aerospike, Inc. under one or more contributor
  * license agreements WHICH ARE COMPATIBLE WITH THE APACHE LICENSE, VERSION 2.0.
@@ -225,7 +225,7 @@ public final class Buffer {
 		int mult = 1;
 
 		for (int i = end - 1; i >= begin; i--) {
-			val += ((int)buf[i] - 48) * mult;
+			val += (buf[i] - 48) * mult;
 			mult *= 10;
 		}
 		return val;
@@ -405,7 +405,7 @@ public final class Buffer {
 	//-------------------------------------------------------
 
 	/**
-	 * Convert long to a 7 byte record version for MRT.
+	 * Convert long to a 7 byte record version for transaction.
 	 */
 	public static void longToVersionBytes(long v, byte[] buf, int offset) {
 		buf[offset++] = (byte)(v >>> 0);
@@ -418,7 +418,7 @@ public final class Buffer {
 	}
 
 	/**
-	 * Convert 7 byte record version to a long for MRT.
+	 * Convert 7 byte record version to a long for transaction.
 	 */
 	public static long versionBytesToLong(byte[] buf, int offset) {
 		return (
